reorganizing and renaming files
