--- examples/loadD4M.jl
@@ -1,3 +1,0 @@
-#A julia script to load D4M if it isn't modularized.
-
-include("../src/D4M.jl")
