+++ src/Assoc/print.jl
@@ -18,6 +18,34 @@ function print(A::Assoc)
     return nothing
 end
 
+#=
+printFull : print Assoc in tabular form.
+=#
+function printFull(A::Assoc)
+    if !isempty(A)
+        display(full(A))
+    else
+        show(A)
+    end
+
+    return nothing
+end
+
+#=
+printTriple : return A in triple String form: (r,c) v 
+=#
+function printTriple(A::Assoc)
+
+    if !isempty(A)
+        r,c,v = find(A)
+        println.("(" .* string.(r) .* "," .* string.(c) .* ")\t".* string.(v))
+    else
+        show(A)
+    end
+
+    return nothing
+end
+
 ########################################################
 # D4M: Dynamic Distributed Dimensional Data Model
 # Architect: Dr. Jeremy Kepner (kepner@ll.mit.edu)
