+++ src/Assoc/put.jl
@@ -24,3 +24,19 @@ function putVal(A::Assoc,AVal::StringOrNum)
     r,c,v = find(A)
     Assoc(r,c,AVal)
 end
+
+#=
+nocol : remove columns mapping
+=#
+
+function nocol(A::Assoc)
+    return condense(Assoc(A.row,promote(ones(size(A.col)),A.col)[1],A.val,A.A))
+end
+
+#=
+norow : remove rows mapping
+=#
+
+function norow(A::Assoc)
+    return condense(Assoc(promote(ones(size(A.row)),A.row)[1],A.col,A.val,A.A))
+end
