--- src/Assoc_orig/emptyAssoc.jl
@@ -1,14 +1,0 @@
-
-#=
-Returns an empty Associative Array
-=#
-function emptyAssoc()
-    Assoc([],[],[])
-end
-
-########################################################
-# D4M: Dynamic Distributed Dimensional Data Model
-# Architect: Dr. Jeremy Kepner (kepner@ll.mit.edu)
-# Software Engineer: Alexander Chen (alexc89@mit.edu)
-########################################################
-
