--- src/Assoc_orig/str2num.jl
@@ -1,7 +1,0 @@
-# Function to convert string values to numeric values
-
-function str2num(A::Assoc)
-    r,c,v = find(A)
-    v = Meta.parse.(v)
-    A = Assoc(r,c,v)
-end
