--- src/NumStr.jl
@@ -1,10 +1,0 @@
-function NumStr(stringA)
-    return length(split(stringA[1:end-1],stringA[end]))
-end
-
-########################################################
-# D4M: Dynamic Distributed Dimensional Data Model
-# Architect: Dr. Jeremy Kepner (kepner@ll.mit.edu)
-# Software Engineer: Alexander Chen (alexc89@mit.edu)
-########################################################
-
