+++ test/runtests.jl
@@ -3,6 +3,8 @@
 using Test, D4M, JLD, PyPlot, SparseArrays
 #include("loadD4M.jl")
 
+import Base.isequal
+
 UnionArray = Array{Union{AbstractString,Number}}
 
 # Check whether two associative arrays are "equal"
@@ -122,55 +124,26 @@ end
         intro_res = load("test_assocs/1intro.jld")["intro_res"]
         @testset "Assoc Intro" begin
             include("../examples/1Intro/1AssocIntro/AI1_SetupTEST.jl")
-            #@test testassoc(A)
             @test testassoc(A,intro_res["AI1_setup"])
+
             include("../examples/1Intro/1AssocIntro/AI2_SubsrefTEST.jl")
-            #@test testassoc(A1r)
             @test testassoc(A1r,intro_res["AI2_subsref"]["A1r"])
-            #@test testassoc(A2r)
-            #@test testassoc(A3r)
             @test testassoc(A3r,intro_res["AI2_subsref"]["A3r"])
-            #@test testassoc(A4r)
             @test testassoc(A4r,intro_res["AI2_subsref"]["A4r"])
-            #@test testassoc(A1c)
             @test testassoc(A1c,intro_res["AI2_subsref"]["A1c"])
-            #@test testassoc(A2c)
-            #@test testassoc(A3c)
             @test testassoc(A3c,intro_res["AI2_subsref"]["A3c"])
-            #@test testassoc(A4c)
             @test testassoc(A4c,intro_res["AI2_subsref"]["A4c"])
-            #@test testassoc(A1v)
+
             include("../examples/1Intro/1AssocIntro/AI3_MathTEST.jl")
-            #@test testassoc(A)
             @test testassoc(A,intro_res["AI3_math"]["A"])
-            #@test testassoc(sum(A,1))
             @test testassoc(sum(A,1),intro_res["AI3_math"]["sum1"])
-            #@test testassoc(sum(A,2))
             @test testassoc(sum(A,2),intro_res["AI3_math"]["sum2"])
-            #@test testassoc(Aab)
             @test testassoc(Aab,intro_res["AI3_math"]["Aab"])
-            #@test testassoc(F)
             @test testassoc(F,intro_res["AI3_math"]["F"])
-            #@test testassoc(Fn)
             @test testassoc(Fn,intro_res["AI3_math"]["Fn"])
-            #@test testassoc(AtA)
             @test testassoc(AtA,intro_res["AI3_math"]["AtA"])
+
             include("../examples/1Intro/1AssocIntro/AI4_AdvConstructTEST.jl")
-            #=
-            @test testassoc(A00)
-            @test testassoc(A01)
-            @test testassoc(A02)
-            @test testassoc(A03)
-            @test testassoc(A04)
-            @test testassoc(A05)
-            @test testassoc(A06)
-            @test testassoc(A11)
-            @test testassoc(A12)
-            @test testassoc(A13)
-            @test testassoc(A14)
-            @test testassoc(A15)
-            @test testassoc(A16)
-            =#
             @test testassoc(A00,intro_res["AI4_construct"]["A00"])
             @test testassoc(A01,intro_res["AI4_construct"]["A01"])
             @test testassoc(A02,intro_res["AI4_construct"]["A02"])
@@ -188,31 +161,15 @@ end
         end
         @testset "Edge Art" begin
             include("../examples/1Intro/2EdgeArt/EA1_GraphTEST.jl")
-            #=
-            @test testassoc(Ev)
-            @test testassoc(Av)
-            @test testassoc(Ae)
-            =#
             @test testassoc(Ev,intro_res["EA1_graph"]["Ev"])
             @test testassoc(Av,intro_res["EA1_graph"]["Av"])
             @test testassoc(Ae,intro_res["EA1_graph"]["Ae"])
             
             include("../examples/1Intro/2EdgeArt/EA2_SubsrefTEST.jl")
-            #=
-            @test testassoc(Eo)
-            @test testassoc(Eog)
-            =#
             @test testassoc(Eo,intro_res["EA2_subsref"]["Eo"])
             @test testassoc(Eog,intro_res["EA2_subsref"]["Eog"])
             
             include("../examples/1Intro/2EdgeArt/EA3_SubGraphTEST.jl")
-            #=
-            @test testassoc(Ev)
-            @test testassoc(EvO)
-            @test testassoc(EvG)
-            @test testassoc(AvOG)
-            #@test testassoc(AeOG)
-            =#
             @test testassoc(EvO,intro_res["EA3_subgraph"]["EvO"])
             @test testassoc(EvG,intro_res["EA3_subgraph"]["EvG"])
             @test testassoc(AvOG,intro_res["EA3_subgraph"]["AvOG"])
@@ -223,30 +180,6 @@ end
     end
     @testset "Apps" begin
         @testset "Entity Analysis" begin
-        #=
-            include("../examples/2Apps/1EntityAnalysis/EA1_ReadTEST.jl")
-            @test testassoc(E)
-            include("../examples/2Apps/1EntityAnalysis/EA2_StatTEST.jl")
-            @test testassoc(sum(logical(col2type(E,"/")),1))
-            @test testassoc(En)
-            @test testassoc(An)
-            include("../examples/2Apps/1EntityAnalysis/EA3_FacetTEST.jl")
-            @test testassoc(F)
-            @test testassoc(Fn)
-            @test testassoc(F' > 1)
-            @test testassoc(Fn' > 0.02)
-            include("../examples/2Apps/1EntityAnalysis/EA4_GraphTEST.jl")
-            @test testassoc(Ae)
-            @test testassoc(Ep)
-            @test testassoc(Ap)
-            @test testassoc(Ad)
-            include("../examples/2Apps/1EntityAnalysis/EA5_GraphQueryTEST.jl")
-            @test testassoc(A)
-            @test testassoc(An)
-            @test testassoc((A[p,x] > 4) & (An[p,x] > 0.3))
-            @test testassoc(A[p1,p1])
-            @test testassoc(A[p2,p2] > 1)
-        =#
             include("../examples/2Apps/1EntityAnalysis/EA1_ReadTEST.jl")
             apps_EA1 = load("test_assocs/2apps_1EA1.jld")["apps_EA1"]
             @test testassoc(E,apps_EA1["E"])
@@ -283,26 +216,6 @@ end
         PyPlot.close()
         PyPlot.close()
         @testset "Track Analysis" begin
-        #=
-            include("../examples/2Apps/2TrackAnalysis/findtrackgraph.jl")
-            include("../examples/2Apps/2TrackAnalysis/findtracks.jl")
-            include("../examples/2Apps/2TrackAnalysis/TA1_BuildTEST.jl")
-            @test testassoc(E3)
-            @test testassoc(Etx)
-            @test testassoc(Ext)
-            @test testassoc(At)
-            @test testassoc(Ax)
-            include("../examples/2Apps/2TrackAnalysis/TA2_QueryTEST.jl")
-            @test testassoc(A)
-            @test testassoc(A[:,p1*p2])
-            @test testassoc(A[t,:] == x)
-            include("../examples/2Apps/2TrackAnalysis/TA3_GraphTEST.jl")
-            @test testassoc(A)
-            @test testassoc(G)
-            @test testassoc(G > 5)
-            @test testassoc(Go)
-            @test testassoc((Go > 2) & ((Go ./ G) > 0.2))
-            =#
             include("../examples/2Apps/2TrackAnalysis/findtrackgraph.jl")
             include("../examples/2Apps/2TrackAnalysis/findtracks.jl")
             include("../examples/2Apps/2TrackAnalysis/TA1_BuildTEST.jl")
@@ -320,11 +233,10 @@ end
             include("../examples/2Apps/2TrackAnalysis/TA3_GraphTEST.jl")
             apps_TA3 = load("test_assocs/2apps_2TA3.jld")["apps_TA3"]
             @test testassoc(A,apps_TA3["A"])
-            @test testassoc(G,apps_TA3["G"
+            @test testassoc(G,apps_TA3["G"])
             @test testassoc(G > 5,apps_TA3["G5"])
             @test testassoc(Go,apps_TA3["Go"])
             @test testassoc((Go > 2) & ((Go ./ G) > 0.2),apps_TA3["GoG"])
-            
         end
     end
     PyPlot.close()
@@ -371,7 +283,6 @@ end
                 println("pDB13_EdgeJoinTEST.jl")
                 include("../examples/3Scaling/2ParallelDatabase/pDB13_EdgeJoinTEST.jl")
                 PyPlot.close()
-
             end
         end
